refactor(fw): Move args to own folder
Moved all files related to arguments and argument handling in commands/
to commands/arguments, and fixed filenames, paths, etc. to clean up
commands/.

undefined
