+++ lib/client.mjs
@@ -7,7 +7,7 @@ import {generateEvents} from './events/generate'
 import {addGroups} from './groups/add'
 import {generateGroups} from './groups/generate'
 import {generatePrefix} from './prefix/generate'
-import {addDefaultTypes} from './commands/add-default-types.mjs'
+import {addDefaultTypes} from './commands/arguments/types/add-defaults'
 
 export class KwClient extends Discord.Client {
   constructor ({events, prefix, groups, commands, info, ...options} = {}) {
