--- lib/commands/add-default-types.mjs
@@ -1,6 +1,0 @@
-import {addTypes} from './add-types'
-import defaultTypes from './default-types'
-
-export function addDefaultTypes () {
-  addTypes(defaultTypes)
-}
