+++ lib/commands/add.mjs
@@ -1,10 +1,9 @@
-import Discord from 'discord.js'
 import tags from 'common-tags'
 import {addEvents} from '../events/add'
 import {validateCommand} from './validate'
-import {ArgumentError} from './argument-error'
-import {getType} from './get-type'
-import {processArgs} from './process-args'
+import {ArgumentError} from './arguments/argument-error'
+import {getType} from './arguments/types/get'
+import {processArgs} from './arguments/process'
 
 export function addCommands (client, commands) {
   if (!Array.isArray(commands)) commands = Object.values(commands)
