+++ lib/commands/arguments/types/add-defaults.mjs
@@ -1,0 +1,6 @@
+import {addTypes} from './add'
+import defaultTypes from './defaults'
+
+export function addDefaultTypes () {
+  addTypes(defaultTypes)
+}
