+++ lib/commands/arguments/types/defaults/boolean.mjs
