+++ lib/commands/arguments/types/defaults/number.mjs
