+++ lib/commands/arguments/types/defaults/string.mjs
