+++ lib/commands/arguments/types/defaults/user.mjs
